Abstract `Configuration` out of `Source::`, add `Configurable` mixin. `Shark::Configuration` is now a top-level type that can be used for more than just Source configurations (i.e., it can be used for Agency configuration as well).
Additionally, `Shark::Configurable` is a mixin module that implements `#configuration`, `#configure`, and `::inherit_configuration_from`, to enable tree-like configuration inheritance through modules, classes, and instances. Simply `include Shark::Configurable` and add `inherit_configuration_from <Configurable object>`, and the object will inherit all configuration attributes from that object.
