+++ sources/doublemap.rb
@@ -4,23 +4,9 @@ require_relative '../source.rb'
 module DoubleMapSource
   extend Shark::Source
 
-
-  class Source
-    # The name of the agency providing information via doublemap. This should
-    # match the subdomain used on doublemap.com
-    attr_accessor :agency
-    # The name of the attribute used to index the data elements maintained by
-    # this source.
-    attr_accessor :key
-    # The data maintained by this source, stored as a hash, indexed by the
-    # primary key attribute given by `key`.
-    attr_accessor :data
-
-    def initialize agency:, key:
-      @agency   = agency
-      @key      = key
-      @data     = {}
-    end
+  class Source < Shark::Source::NormalSource
+    include Shark::Configurable
+    inherit_configuration_from DoubleMapSource
 
     def doublemap
       @doublemap ||= begin
@@ -33,6 +19,8 @@ module DoubleMapSource
       end
     end
   end
+
+  register_source :doublemap, Shark::Route, DoubleMapSource::Source
 end
 
 # require_relative 'doublemap/route'
